+++ src/config.php
@@ -140,6 +140,7 @@ return [
         ],
         'strictTransportSecurity' => [
             'canFail' => true,
+            'threshold' => 15552000, // 6 months
         ],
         'xContentTypeOptions' => [
             'canFail' => true,
@@ -197,6 +198,9 @@ return [
         'requireUserAgentAndIpForSession' => [],
         'sanitizeSvgUploads' => [
             'canFail' => true,
+        ],
+        'sendPoweredByHeader' => [
+            'canFail' => false,
         ],
         'testToEmailAddress' => [
             'canFail' => true,